Update README.md
Hindi numerals, also known as Indian numerals, are the symbols used to represent numbers in the Hindi language. These numerals are commonly used in the Indian subcontinent and are distinct from the Arabic numerals used in Western countries. HindiNumerals gem aims to facilitate the conversion of Arabic numbers to their Hindi equivalents, allowing developers to seamlessly work with Hindi numerals in their Ruby applications.